✨(frontend) update organization form
Transform the organization form by separating it into several tabs
