+++ src/frontend/admin/src/components/templates/organizations/form/OrganizationForm.tsx
@@ -1,50 +1,50 @@
 import * as React from "react";
-import { useEffect } from "react";
-import * as Yup from "yup";
-import { useForm } from "react-hook-form";
-import { yupResolver } from "@hookform/resolvers/yup";
-import Grid from "@mui/material/Unstable_Grid2";
-import { FormattedMessage, useIntl } from "react-intl";
-import Box from "@mui/material/Box";
-import Stack from "@mui/material/Stack";
-import Typography from "@mui/material/Typography";
-import Alert from "@mui/material/Alert";
-import { RHFProvider } from "@/components/presentational/hook-form/RHFProvider";
-import { RHFTextField } from "@/components/presentational/hook-form/RHFTextField";
-import { RHFUploadImage } from "@/components/presentational/hook-form/RHFUploadImage";
-import {
-  DTOOrganization,
-  Organization,
-  OrganizationRoles,
-} from "@/services/api/models/Organization";
-import { commonTranslations } from "@/translations/common/commonTranslations";
-import { organizationFormMessages } from "@/components/templates/organizations/form/translations";
-import { useOrganizations } from "@/hooks/useOrganizations/useOrganizations";
-import { Maybe, ServerSideErrorForm } from "@/types/utils";
-import { genericUpdateFormError } from "@/utils/forms";
-import { TranslatableContent } from "@/components/presentational/translatable-content/TranslatableContent";
-import { SimpleCard } from "@/components/presentational/card/SimpleCard";
-import { AccessesList } from "@/components/templates/accesses/list/AccessesList";
-import { LoadingContent } from "@/components/presentational/loading/LoadingContent";
-import { ThumbnailDetailField } from "@/services/api/models/Image";
-import { RHFSelect } from "@/components/presentational/hook-form/RHFSelect";
+import { useMemo } from "react";
+import { defineMessages, useIntl } from "react-intl";
+import { Organization } from "@/services/api/models/Organization";
 import { OrganizationAddressForm } from "@/components/templates/organizations/form/sections/OrganizationAddressForm";
+import { OrganizationGeneralSection } from "@/components/templates/organizations/form/sections/OrganizationGeneralSection";
+import { OrganizationFormMemberSection } from "@/components/templates/organizations/form/sections/OrganizationMemberSection";
+import {
+  TabsComponent,
+  TabValue,
+} from "@/components/presentational/tabs/TabsComponent";
 
-interface FormValues {
-  code: string;
-  title: string;
-  representative: string | undefined;
-  signature: File[] | undefined;
-  logo: File[] | undefined;
-  enterprise_code?: string; // SIRET in France
-  activity_category_code?: string; // APE in France
-  representative_profession?: string;
-  signatory_representative?: string;
-  signatory_representative_profession?: string;
-  contact_phone?: string;
-  contact_email?: string;
-  dpo_email?: string;
-}
+const messages = defineMessages({
+  generalTabTitle: {
+    id: "components.templates.organizations.address.form.translations.generalTabTitle",
+    defaultMessage: "General",
+    description: "Title for general tab",
+  },
+  addressTabTitle: {
+    id: "components.templates.organizations.address.form.translations.addressTabTitle",
+    defaultMessage: "Address",
+    description: "Title for address tab",
+  },
+  memberTabTitle: {
+    id: "components.templates.organizations.address.form.translations.memberTabTitle",
+    defaultMessage: "Members",
+    description: "Title for members tab",
+  },
+  generalTabInfo: {
+    id: "components.templates.organizations.address.form.translations.generalTabInfo",
+    defaultMessage:
+      "In this section, you can provide general information, give information on the signatories or even on the legal part of the organization",
+    description: "Text for general tab info",
+  },
+  addressTabInfo: {
+    id: "components.templates.organizations.address.form.translations.addressTabInfo",
+    defaultMessage:
+      "In this section, you must fill in the details of the organization's address",
+    description: "Text for address tab info",
+  },
+  memberTabInfo: {
+    id: "components.templates.organizations.address.form.translations.memberTabInfo",
+    defaultMessage:
+      "In this section, you can manage the members of the organization and their roles. Please note, there must be at least one owner of the organization",
+    description: "Text for members tab info",
+  },
+});
 
 interface Props {
   afterSubmit?: (values: Organization) => void;
@@ -54,349 +54,43 @@ interface Props {
 
 export function OrganizationForm(props: Props) {
   const intl = useIntl();
-  const organizationQuery = useOrganizations({}, { enabled: false });
   const defaultOrganization = props.organization ?? props.fromOrganization;
-
-  const getDefaultValues = () => {
-    return {
-      code: defaultOrganization?.code ?? "",
-      title: defaultOrganization?.title ?? "",
-      representative: defaultOrganization?.representative ?? "",
-      country: defaultOrganization?.country ?? "FR",
-      enterprise_code: defaultOrganization?.enterprise_code ?? "",
-      activity_category_code: defaultOrganization?.activity_category_code ?? "",
-      representative_profession:
-        defaultOrganization?.representative_profession ?? "",
-      signatory_representative:
-        defaultOrganization?.signatory_representative ?? "",
-      signatory_representative_profession:
-        defaultOrganization?.signatory_representative_profession ?? "",
-      contact_phone: defaultOrganization?.contact_phone ?? "",
-      contact_email: defaultOrganization?.contact_email ?? "",
-      dpo_email: defaultOrganization?.dpo_email ?? "",
-    };
-  };
-
-  const RegisterSchema = Yup.object().shape({
-    code: Yup.string().required(),
-    title: Yup.string().required(),
-    representative: Yup.string(),
-    signature: Yup.mixed(),
-    logo: Yup.mixed(),
-    country: Yup.string(),
-    enterprise_code: Yup.string(),
-    activity_category_code: Yup.string(),
-    representative_profession: Yup.string(),
-    signatory_representative: Yup.string(),
-    signatory_representative_profession: Yup.string().when(
-      ["signatory_representative"],
+  const tabs = useMemo(() => {
+    let result: TabValue[] = [
       {
-        is: (signatory_representative: Maybe<string>) =>
-          signatory_representative != null && signatory_representative !== "",
-        then: (schema) => schema.required(),
+        label: intl.formatMessage(messages.generalTabTitle),
+        tabInfo: intl.formatMessage(messages.generalTabInfo),
+        component: (
+          <OrganizationGeneralSection
+            organization={defaultOrganization}
+            afterSubmit={props.afterSubmit}
+          />
+        ),
       },
-    ),
-    contact_phone: Yup.string(),
-    contact_email: Yup.string().email(),
-    dpo_email: Yup.string().email(),
-  });
-
-  const methods = useForm({
-    resolver: yupResolver(RegisterSchema),
-    defaultValues: getDefaultValues() as any, // To not trigger type validation for default value
-  });
-
-  const updateFormError = (errors: ServerSideErrorForm<FormValues>) => {
-    genericUpdateFormError(errors, methods.setError);
-  };
-
-  const onSubmit = (values: FormValues): void => {
-    const payload: DTOOrganization = {
-      ...values,
-      logo: values.logo?.[0],
-      signature: values.signature?.[0],
-    };
-
-    if (values.logo?.[0] === undefined) {
-      delete payload.logo;
-    }
-    if (values.signature?.[0] === undefined) {
-      delete payload.signature;
-    }
+    ];
 
     if (props.organization) {
-      payload.id = props.organization.id;
-      organizationQuery.methods.update(payload, {
-        onSuccess: (data) => props.afterSubmit?.(data),
-        onError: (error) => updateFormError(error.data),
-      });
-    } else {
-      organizationQuery.methods.create(payload, {
-        onSuccess: (data) => props.afterSubmit?.(data),
-        onError: (error) => updateFormError(error.data),
-      });
+      result = [
+        ...result,
+        {
+          label: intl.formatMessage(messages.addressTabTitle),
+          tabInfo: intl.formatMessage(messages.addressTabInfo),
+          show: !!defaultOrganization,
+          component: (
+            <OrganizationAddressForm organization={props.organization} />
+          ),
+        },
+        {
+          label: intl.formatMessage(messages.memberTabTitle),
+          tabInfo: intl.formatMessage(messages.memberTabInfo),
+          component: (
+            <OrganizationFormMemberSection organization={props.organization} />
+          ),
+        },
+      ];
     }
-  };
+    return result;
+  }, [props.organization, defaultOrganization, intl]);
 
-  const getUploadedSignature = (): ThumbnailDetailField[] => {
-    if (props.fromOrganization || !organizationQuery) {
-      return [];
-    }
-    return defaultOrganization?.signature
-      ? [defaultOrganization.signature]
-      : [];
-  };
-
-  const getUploadedLogo = (): ThumbnailDetailField[] => {
-    if (props.fromOrganization || !organizationQuery) {
-      return [];
-    }
-    return defaultOrganization?.logo ? [defaultOrganization.logo] : [];
-  };
-
-  useEffect(() => {
-    methods.reset(getDefaultValues());
-  }, [props.organization]);
-
-  const { handleSubmit } = methods;
-  return (
-    <Stack spacing={3}>
-      <SimpleCard>
-        <TranslatableContent
-          onSelectLang={() => {
-            if (props.organization) organizationQuery.methods.invalidate();
-          }}
-        >
-          <Box padding={4}>
-            <RHFProvider
-              checkBeforeUnload={true}
-              showSubmit={true}
-              methods={methods}
-              isSubmitting={organizationQuery.states.updating}
-              id="organization-form"
-              onSubmit={handleSubmit(onSubmit)}
-            >
-              <Grid container spacing={2}>
-                <Grid xs={12}>
-                  <Typography variant="subtitle2">
-                    <FormattedMessage
-                      {...organizationFormMessages.generalSectionTitle}
-                    />
-                  </Typography>
-                </Grid>
-
-                <Grid xs={12} md={6}>
-                  <RHFTextField
-                    name="title"
-                    label={intl.formatMessage(commonTranslations.title)}
-                  />
-                </Grid>
-                <Grid xs={12} md={6}>
-                  <RHFTextField
-                    name="code"
-                    label={intl.formatMessage(
-                      organizationFormMessages.codeLabel,
-                    )}
-                  />
-                </Grid>
-
-                <Grid xs={12} md={6}>
-                  <RHFTextField
-                    name="representative"
-                    label={intl.formatMessage(
-                      organizationFormMessages.representativeLabel,
-                    )}
-                  />
-                </Grid>
-                <Grid xs={12} md={6}>
-                  <RHFTextField
-                    name="representative_profession"
-                    label={intl.formatMessage(
-                      organizationFormMessages.representativeProfessionLabel,
-                    )}
-                  />
-                </Grid>
-
-                <Grid xs={12}>
-                  <RHFSelect
-                    disabled={!organizationQuery.countries}
-                    name="country"
-                    options={organizationQuery.countries ?? []}
-                    label={intl.formatMessage(
-                      organizationFormMessages.countryLabel,
-                    )}
-                  />
-                </Grid>
-
-                <Grid xs={12}>
-                  <RHFUploadImage
-                    thumbnailFiles={getUploadedLogo()}
-                    buttonLabel={intl.formatMessage(
-                      organizationFormMessages.uploadLogoButtonLabel,
-                    )}
-                    name="logo"
-                    accept="image/*"
-                    label={intl.formatMessage(
-                      organizationFormMessages.logoLabel,
-                    )}
-                  />
-                </Grid>
-
-                <Grid xs={12}>
-                  <Typography variant="subtitle2">
-                    <FormattedMessage
-                      {...organizationFormMessages.signatoryDetailsSectionTitle}
-                    />
-                  </Typography>
-                </Grid>
-
-                <Grid xs={12}>
-                  <Alert severity="info">
-                    <FormattedMessage
-                      {...organizationFormMessages.signatoryDetailsSectionInfo}
-                    />
-                  </Alert>
-                </Grid>
-
-                <Grid xs={12} md={6}>
-                  <RHFTextField
-                    name="signatory_representative"
-                    label={intl.formatMessage(
-                      organizationFormMessages.signatoryRepresentativeLabel,
-                    )}
-                  />
-                </Grid>
-
-                <Grid xs={12} md={6}>
-                  <RHFTextField
-                    name="signatory_representative_profession"
-                    helperText={intl.formatMessage(
-                      organizationFormMessages.signatoryRepresentativeProfessionHelperText,
-                    )}
-                    label={intl.formatMessage(
-                      organizationFormMessages.signatoryRepresentativeProfessionLabel,
-                    )}
-                  />
-                </Grid>
-                <Grid xs={12}>
-                  <RHFUploadImage
-                    thumbnailFiles={getUploadedSignature()}
-                    name="signature"
-                    buttonLabel={intl.formatMessage(
-                      organizationFormMessages.uploadSignatureButtonLabel,
-                    )}
-                    accept="image/*"
-                    label={intl.formatMessage(
-                      organizationFormMessages.signatureLabel,
-                    )}
-                  />
-                </Grid>
-
-                <Grid xs={12}>
-                  <Typography variant="subtitle2">
-                    <FormattedMessage
-                      {...organizationFormMessages.legalPartSectionTitle}
-                    />
-                  </Typography>
-                </Grid>
-                <Grid xs={12} md={6}>
-                  <RHFTextField
-                    name="enterprise_code"
-                    label={intl.formatMessage(
-                      organizationFormMessages.enterpriseCodeLabel,
-                    )}
-                  />
-                </Grid>
-                <Grid xs={12} lg={6}>
-                  <RHFTextField
-                    name="activity_category_code"
-                    label={intl.formatMessage(
-                      organizationFormMessages.activityCategoryCodeLabel,
-                    )}
-                  />
-                </Grid>
-
-                <Grid xs={12}>
-                  <Typography variant="subtitle2">
-                    <FormattedMessage
-                      {...organizationFormMessages.contactSectionTitle}
-                    />
-                  </Typography>
-                </Grid>
-                <Grid xs={12} md={6} lg={4}>
-                  <RHFTextField
-                    name="contact_phone"
-                    label={intl.formatMessage(
-                      organizationFormMessages.contactPhoneLabel,
-                    )}
-                  />
-                </Grid>
-                <Grid xs={12} md={6} lg={4}>
-                  <RHFTextField
-                    name="contact_email"
-                    label={intl.formatMessage(
-                      organizationFormMessages.contactEmailLabel,
-                    )}
-                  />
-                </Grid>
-                <Grid xs={12} lg={4}>
-                  <RHFTextField
-                    name="dpo_email"
-                    label={intl.formatMessage(
-                      organizationFormMessages.dpoContactEmailLabel,
-                    )}
-                  />
-                </Grid>
-              </Grid>
-            </RHFProvider>
-          </Box>
-        </TranslatableContent>
-      </SimpleCard>
-      {defaultOrganization && (
-        <OrganizationAddressForm organization={defaultOrganization} />
-      )}
-      <LoadingContent loading={organizationQuery.accesses === undefined}>
-        {props.organization && organizationQuery.accesses && (
-          <SimpleCard>
-            <Box padding={4}>
-              <Typography>
-                {intl.formatMessage(
-                  organizationFormMessages.membersSectionTitle,
-                )}
-              </Typography>
-            </Box>
-            <AccessesList
-              defaultRole={OrganizationRoles.MEMBER}
-              onRemove={async (accessId) => {
-                await organizationQuery.methods.removeAccessUser(
-                  // @ts-ignore
-                  props.organization?.id,
-                  accessId,
-                );
-              }}
-              onUpdateAccess={(accessId, payload) => {
-                return organizationQuery.methods.updateAccessUser(
-                  // @ts-ignore
-                  props.organization.id,
-                  accessId,
-                  payload,
-                );
-              }}
-              onAdd={(user, role) => {
-                if (props.organization?.id && user.id) {
-                  organizationQuery.methods.addAccessUser(
-                    props.organization?.id,
-                    user.id,
-                    role,
-                  );
-                }
-              }}
-              accesses={props.organization?.accesses ?? []}
-              availableAccesses={organizationQuery.accesses}
-            />
-          </SimpleCard>
-        )}
-      </LoadingContent>
-    </Stack>
-  );
+  return <TabsComponent id="organization-form-tabs" tabs={tabs} />;
 }
